add consumer tests

+++ programs/aquachain/src/instructions/register_consumer.rs
@@ -48,13 +48,13 @@ pub fn register_consumer(
     tariff_key: Pubkey,
     reservoir_key: Pubkey,
     contracted_capacity: u64,
-    block_rate: f64,
+    block_rate: u64,
 ) -> Result<()> {
     let consumer = &mut ctx.accounts.consumer;
 
     // Validation: Ensure capacity and rate are non-zero
     require!(contracted_capacity > 0, CustomError::InvalidCapacity);
-    require!(block_rate > 0.0, CustomError::InvalidRate);
+    require!(block_rate > 0, CustomError::InvalidRate);
 
     consumer.assigned_tariff = tariff_key;
     consumer.assigned_reservoir = reservoir_key;
