+++ programs/aquachain/src/state/consumer.rs
@@ -3,7 +3,7 @@ use anchor_lang::prelude::*;
 #[account]
 #[derive(InitSpace)]
 pub struct Consumer {
-    pub block_rate: f64,
+    pub block_rate: u64,
     pub contracted_capacity: u64,
     pub assigned_tariff: Pubkey,
     pub assigned_reservoir: Pubkey,
